Refactor cuanto.parsing package
Spilt the mix-in to create a complete parser from enriching Parsers.
